Deep Learning Application in Computer Vision for Traffic Sign Classification and Detection
In summary, this project seeks to explore the science behind Neural
Networks (NN), its various flavours, application areas and then finally,
narrow down by applying it in the design and development of a computer
vision system which can be used for traffic sign recognition and
detection in autonomous vehicles. The project starts off by designing,
developing, implementing and testing a model of the proposed vision
system on a CPU using MATLAB and then afterwards, the performance of the
implemented vision system is further optimized through vectorization,
parallelism, legacy coding and heterogeneous computing. The project is
concluded with detailed analysis and evaluation of the various
optimization schemes utilized as well as an evaluation of the excellent
Neural Network’s classification accuracy.